// src/app/api/models/poll-content.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MessageContent } from '../models/message-content';

/**
 * Poll message content
 */
export type PollContent = MessageContent & {

/**
 * Poll question
 */
'question'?: string;

/**
 * Poll options
 */
'options'?: Array<string>;

/**
 * Whether multiple selections are allowed
 */
'allowMultipleSelections'?: boolean;
} & {
'$type': 'poll';

/**
 * Type discriminator (always 'poll' for PollContent)
 */
'$type': string;
};